+++ Changelog.md
@@ -1,6 +1,19 @@
 # Change Log
 
 
+## Unreleased future version
+
+New
+
+* Status information is now available in JSON format at `/api/v1/status`
+  ([#449]).
+* Includes version 0.1.0 of [routinator-ui], a UI for Route Origin Validation
+  and Routinator status ([#449]).
+
+[#449]: https://github.com/NLnetLabs/routinator/pull/449
+[routinator-ui]: https://github.com/NLnetLabs/routinator-ui/
+
+
 ## 0.8.2 ‘Once More, with Feeling’
 
 Released 2020-12-09.